fix mingw compile add Ws2tcpip.h in sys/socket to use some more defines from winapi instead our own ones
svn path=/trunk/KDE/kdelibs/; revision=509433

+++ win/src/unistd.c
@@ -146,7 +146,7 @@ typedef unsigned int size_t;
 
 KDEWIN32_EXPORT int kde_gethostname(char *__name, size_t __len)
 {
-  size_t len = __len;
+  DWORD len = __len;
   if (0==GetComputerNameA(__name, &len))
     return -1;
   return 0;
